Moving kernels involving physics tendancy variables on to the CPU for CAM Dycore
The physics tendancy variables were changed from an allocatable to
a pointer variable. This is done to accomodate the variables coming
from CAM. These module pointer variables causes issues with OpenACC
execution. As a workaround, currently the kernels involving
tend variables are moved to CPU for CAM dycore execution. The issue
has been communicated to the NVHPC compiler group.